initial clean up and bug fixes

+++ database/migrations/2024_09_05_072833_create_tasks_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('tasks', function (Blueprint $table) {
             $table->id();
             $table->string('description');
-            $table->unsignedBigInteger('estimated_time');
-            $table->unsignedBigInteger('used_time')->default(0);
+            $table->float('estimated_time');
+            $table->float('used_time')->default(0);
             $table->date('created_date');
             $table->date('completed_date')->nullable()->default(null);
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
